Fix NewWindow3 API workaround
Changes the workaround to obtain the new window handle asynchronously.
Without this change, the new window handle was not available 50%
of the time, at least on some computers.

In addition, now the browser UI is not blocked several seconds
when a new window/tab is opened.
Previously it was sometimes blocked for up to 10 seconds.